pyproject.toml
[project]
name = "hybrid-resume-screening-pipeline"
version = "0.1.0"
description = "Prototype of hybrid resume screening pipeline"
authors = [
    {name = "David Adediji",email = "david.adediji@sil.mu"}
]
readme = "README.md"
requires-python = ">=3.9, <3.12"
dependencies = [
    "openai (>=2.3.0,<3.0.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.104.0,<1.0.0)",
    "uvicorn[standard] (>=0.24.0,<1.0.0)",
    "python-multipart (>=0.0.6,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
